correctly support 1.19+ & use gradle
