--- 1_15_2/pom.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>RPGHorses</artifactId>
-        <groupId>org.plugins.rpghorses</groupId>
-        <version>1.0.0</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>1_15_2</artifactId>
-    <groupId>org.plugins.rpghorses</groupId>
-    <packaging>jar</packaging>
-    <version>1.0.0</version>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.plugins.rpghorses</groupId>
-            <artifactId>API</artifactId>
-            <version>1.0.0</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>local.spigotmc</groupId>
-            <artifactId>spigot</artifactId>
-            <version>1.15.2</version>
-            <scope>provided</scope>
-            <optional>true</optional>
-        </dependency>
-    </dependencies>
-
-</project>
