+++ API/build.gradle
@@ -1,0 +1,5 @@
+dependencies {
+    compileOnly 'org.spigotmc:spigot-api:1.19.3-R0.1-SNAPSHOT'
+}
+
+description = 'API'
