+++ NMS/1_11_2/src/main/java/org/plugins/rpghorses/v1_11_2/NMSHandler.java
