+++ NMS/1_13_2/build.gradle
@@ -1,0 +1,6 @@
+dependencies {
+    compileOnly project(':API')
+    compileOnly 'org.spigotmc:spigot:1.13.2-R0.1-SNAPSHOT'
+}
+
+description = '1_13_2'
