+++ NMS/1_14_4/build.gradle
@@ -1,0 +1,6 @@
+dependencies {
+    compileOnly project(':API')
+    compileOnly 'org.spigotmc:spigot:1.14.4-R0.1-SNAPSHOT'
+}
+
+description = '1_14_4'
