+++ NMS/1_16_5/build.gradle
@@ -1,0 +1,6 @@
+dependencies {
+    compileOnly project(':API')
+    compileOnly 'org.spigotmc:spigot:1.16.5-R0.1-SNAPSHOT'
+}
+
+description = '1_16_5'
