+++ NMS/1_17/build.gradle
@@ -1,0 +1,7 @@
+dependencies {
+    compileOnly project(':API')
+    compileOnly 'org.spigotmc:spigot:1.17-R0.1-SNAPSHOT'
+    compileOnly 'org.spigotmc:spigot-api:1.17-R0.1-SNAPSHOT'
+}
+
+description = '1_17'
