+++ NMS/1_19_3/build.gradle
@@ -1,0 +1,6 @@
+dependencies {
+    compileOnly project(':API')
+    compileOnly 'org.spigotmc:spigot:1.19.3-R0.1-SNAPSHOT'
+}
+
+description = '1_19_3'
