+++ RPGHorsesPlugin/src/main/java/org/plugins/rpghorses/version/Version.java
@@ -22,7 +22,10 @@ public enum Version {
 	v1_18("1.18", 21, "v1_18"),
 	v1_18_1("1.18.1", 22, "v1_18"),
 	v1_18_2("1.18.2", 23, "v1_18_2"),
-	v1_19("1.19", 24, "v1_19"),
+	v1_19("1.19", 24, "v1_19_2"),
+	v1_19_1("1.19.1", 25, "v1_19_2"),
+	v1_19_2("1.19.2", 26, "v1_19_2"),
+	v1_19_3("1.19.3", 27, "v1_19_3"),
 	LATEST("", 100, "");
 	
 	private String name;
