--- pom.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <groupId>org.plugins.rpghorses</groupId>
-    <artifactId>RPGHorses</artifactId>
-    <version>1.0.0</version>
-    <modules>
-        <module>1_11_2</module>
-        <module>1_12_2</module>
-        <module>1_13_2</module>
-        <module>1_14_4</module>
-        <module>1_15_2</module>
-        <module>1_16_3</module>
-        <module>1_16_4</module>
-        <module>1_16_5</module>
-        <module>1_17</module>
-        <module>1_18_1</module>
-        <module>1_18_2</module>
-        <module>API</module>
-        <module>RPGHorsesPlugin</module>
-        <module>1_19</module>
-    </modules>
-    <packaging>pom</packaging>
-
-    <name>RPGHorses</name>
-
-    <description>Adds RPG-like horses to the game</description>
-    <properties>
-        <java.version>1.8</java.version>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    </properties>
-
-    <build>
-        <defaultGoal>clean package</defaultGoal>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.8.1</version>
-                <configuration>
-                    <source>${java.version}</source>
-                    <target>${java.version}</target>
-                </configuration>
-            </plugin>
-        </plugins>
-        <resources>
-            <resource>
-                <directory>src/main/resources</directory>
-                <filtering>true</filtering>
-            </resource>
-        </resources>
-    </build>
-
-
-</project>
