+++ settings.gradle
@@ -1,0 +1,32 @@
+rootProject.name = 'RPGHorses'
+
+include(':RPGHorsesPlugin')
+include(':API')
+
+include(':1_11_2')
+include(':1_12_2')
+include(':1_13_2')
+include(':1_14_4')
+include(':1_15_2')
+include(':1_16_3')
+include(':1_16_4')
+include(':1_16_5')
+include(':1_17')
+include(':1_18_1')
+include(':1_18_2')
+include(':1_19_2')
+include(':1_19_3')
+
+project(":1_11_2").projectDir = file('NMS/1_11_2')
+project(":1_12_2").projectDir = file('NMS/1_12_2')
+project(":1_13_2").projectDir = file('NMS/1_13_2')
+project(":1_14_4").projectDir = file('NMS/1_14_4')
+project(":1_15_2").projectDir = file('NMS/1_15_2')
+project(":1_16_3").projectDir = file('NMS/1_16_3')
+project(":1_16_4").projectDir = file('NMS/1_16_4')
+project(":1_16_5").projectDir = file('NMS/1_16_5')
+project(":1_17").projectDir = file('NMS/1_17')
+project(":1_18_1").projectDir = file('NMS/1_18_1')
+project(":1_18_2").projectDir = file('NMS/1_18_2')
+project(":1_19_2").projectDir = file('NMS/1_19_2')
+project(":1_19_3").projectDir = file('NMS/1_19_3')
